+++ hydra-node/test/Hydra/Chain/Direct/ContractSpec.hs
@@ -19,6 +19,7 @@ import Hydra.Cardano.Api (
   toLedgerTxOut,
   toPlutusTxOut,
  )
+import Hydra.Cardano.Api.Network (networkIdToNetwork)
 import Hydra.Chain.Direct.Contract.Abort (genAbortMutation, healthyAbortTx, propHasCommit, propHasInitial)
 import Hydra.Chain.Direct.Contract.Close (genCloseInitialMutation, genCloseMutation, healthyCloseInitialTx, healthyCloseTx)
 import Hydra.Chain.Direct.Contract.CollectCom (genCollectComMutation, healthyCollectComTx)
@@ -27,6 +28,7 @@ import Hydra.Chain.Direct.Contract.Contest (genContestMutation, healthyContestTx
 import Hydra.Chain.Direct.Contract.FanOut (genFanoutMutation, healthyFanoutTx)
 import Hydra.Chain.Direct.Contract.Init (genInitMutation, healthyInitTx)
 import Hydra.Chain.Direct.Contract.Mutation (propMutation, propTransactionEvaluates)
+import Hydra.Chain.Direct.Fixture (testNetworkId)
 import qualified Hydra.Contract.Commit as Commit
 import Hydra.Contract.Head (
   verifyPartySignature,
@@ -140,7 +142,7 @@ prop_serializingCommitRoundtrip =
   -- Plutus.TxOut
   forAll ((,) <$> arbitrary <*> (arbitrary >>= genOutput)) $ \singleUTxO ->
     let serialized = Commit.serializeCommit singleUTxO
-        deserialized = serialized >>= Commit.deserializeCommit
+        deserialized = serialized >>= Commit.deserializeCommit (networkIdToNetwork testNetworkId)
      in case deserialized of
           Just actual -> actual === singleUTxO
           Nothing ->